% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalquartet.R
\docType{data}
\name{causalquartet}
\alias{causalquartet}
\title{Causal Quartet Data}
\format{
A dataframe with 100 rows and 6 variables:

* `u1`: an unknown factor
* `u2`: an unknown factor
* `z`: a known factor
* `x`: exposure
* `y`: outcome
* `type`: The data generating mechanism
}
\usage{
causalquartet
}
\description{
This dataset contains 400 observations, each generated under a different data generating mechanism
}
\keyword{datasets}
