% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_funcs.R
\name{ColSums}
\alias{ColSums}
\title{Column Sums
An accellerated version of the R function \code{colSums(x, na.rm = FALSE, dims = 1L)}}
\usage{
ColSums(x, n_cols = ncol(x))
}
\arguments{
\item{x}{Matrix whose columns are to be summed.}

\item{n_cols}{Number of columns in said matrix.}
}
\description{
Column Sums
An accellerated version of the R function \code{colSums(x, na.rm = FALSE, dims = 1L)}
}
\author{
Martin R. Smith
}
\keyword{internal}
