% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confounding.R
\docType{data}
\name{confounding}
\alias{confounding}
\title{Confounder Data}
\format{
A dataframe with 100 rows and 3:

* `z`: a known factor (confounder)
* `x`: exposure
* `y`: outcome
}
\usage{
confounding
}
\description{
This dataset contains 100 observations, generated under the following mechanism:
Z ~ N(0, 1)
X ~ Z + N(0,1)
Y ~ 0.5X + Z + N(0, 1)
}
\keyword{datasets}
