% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpdt_dist_file}
\alias{cpdt_dist_file}
\title{Direct entry points to 'cpdt' functions}
\usage{
cpdt_dist_file(file1, file2)
}
\arguments{
\item{file1, file2}{Paths to files containing a tree or trees in Newick format.}
}
\value{
The distance between the requested trees.
}
\description{
Functions to calculate rooted triplet distances between pairs of trees.
Input is not checked for sanity.
}
\details{
Functions are called from R with functions such as \code{\link{CPDTDist}}.
}
\references{
\insertRef{Jansson2017jcb}{Quartet}
}
\author{
Martin R. Smith, after Ramesh Rajaby
}
\keyword{internal}
