% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metrics.R
\name{.AddSlice}
\alias{.AddSlice}
\title{Add slice to 3D array}
\usage{
.AddSlice(arr, slice, sliceName = NULL)
}
\arguments{
\item{arr}{Three-dimensional array.}

\item{slice}{Two-dimensional matrix to add to array.}

\item{sliceName}{Character vector specifying name for new slice.}
}
\value{
A three-dimensional array formed by adding \code{slice} to the end of
\code{arr}.
}
\description{
Add slice to 3D array
}
\author{
\href{https://orcid.org/0000-0001-5660-1727}{Martin R. Smith}
(\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
\keyword{internal}
