% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tqDist.r
\name{.CheckSize}
\alias{.CheckSize}
\title{Check tree size}
\usage{
.CheckSize(tree)
}
\description{
Trees with > 477 leaves may have counts > .Machine$integer.max so cannot
be reliably evaluated.
}
\details{
It may be possible to increase this number to 568 by converting what R
represents as negative integers to the unsigned equivalent that is sent
from C.
}
\keyword{internal}
