% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metrics.R
\name{.NormalizeStatus}
\alias{.NormalizeStatus}
\title{Normalize element statuses to generate metric}
\usage{
.NormalizeStatus(elementStatus, numerator, denominator, takeFromOne)
}
\arguments{
\item{elementStatus}{Two-dimensional integer array, with rows corresponding to
counts of matching quartets or partitions for each tree, and columns named
according to the output of \code{[QuartetStatus]} or
\code{[SplitStatus]}.}

\item{numerator, denominator}{Character vector listing elements to sum in
numerator / denominator}

\item{takeFromOne}{Logical specifying whether to deduct value from one.}
}
\description{
Handles vectors and matrices of two or three dimensions.
}
\keyword{internal}
