% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpdt.R
\name{.PairsTripletDist}
\alias{.PairsTripletDist}
\title{Pairwise triplet distances between corresponding trees}
\usage{
.PairsTripletDist(trees1, trees2)
}
\description{
Pairwise triplet distances between corresponding trees
}
\keyword{internal}
