% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuartetRogue.R
\name{.QuartetRogueScoreSequence}
\alias{.QuartetRogueScoreSequence}
\title{Compute score sequence for a given drop order}
\usage{
.QuartetRogueScoreSequence(trees, dropped)
}
\arguments{
\item{trees}{Original trees.}

\item{dropped}{Character vector of taxa in drop order.}
}
\value{
Numeric vector of scores (length = \code{length(dropped) + 1}).
}
\description{
Compute score sequence for a given drop order
}
\keyword{internal}
