% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareQuartets.R
\name{.Subsplit4}
\alias{.Subsplit4}
\title{Closest to tip 1 in each split}
\usage{
.Subsplit4(tips, splits, nTip = NTip(splits)[1])
}
\arguments{
\item{tips}{Vector of length four specifying index of tips to consider.}

\item{splits}{Splits object.}

\item{nTip}{Integer specifying number of splits in \code{splits}.}
}
\value{
Raw vector specifying the closest relative of \code{tips[1]} in each
split
}
\description{
Closest to tip 1 in each split
}
\keyword{internal}
