% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICQ.R
\name{ICQ}
\alias{ICQ}
\title{Non-redundant quartet distance}
\usage{
ICQ(tr1, tr2, similarity = FALSE, states = NULL, sampleSize = 1000L)
}
\description{
Draft implementation by explicit enumeration across all tree topologies (!)
}
\details{
Use on small trees only!
}
\examples{
library('TreeTools', quietly = TRUE, warn.conflicts = FALSE)
tr1 <- as.phylo(0, 5)
tr2 <- as.phylo(6, 5)

ICQ(tr1, tr1)
ICQ(tr1, tr2)

}
\author{
\href{https://orcid.org/0000-0001-5660-1727}{Martin R. Smith}
(\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
