% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediator.R
\docType{data}
\name{mediator}
\alias{mediator}
\title{Mediator Data}
\format{
A dataframe with 100 rows and 3 variables:

* `x`: exposure
* `z`: a known factor (mediator)
* `y`: outcome
}
\usage{
mediator
}
\description{
This dataset contains 100 observations, generated under the following mechanism:
X ~ N(0, 1)
Z ~ X + N(0,1)
Y ~ Z + N(0, 1)
}
\keyword{datasets}
