% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_funcs.R
\name{NumberTips}
\alias{NumberTips}
\title{Number Tips
Renumber the tips of a tree to match a list of labels}
\usage{
NumberTips(tr, sorted.labels)
}
\arguments{
\item{tr}{A tree of class phylo.}

\item{sorted.labels}{A character vector listing tip labels in the desired order}
}
\description{
Number Tips
Renumber the tips of a tree to match a list of labels
}
\author{
Martin R. Smith
}
\keyword{internal}
