% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PartitionDistance.R
\name{PairSharedSplitStatus}
\alias{PairSharedSplitStatus}
\title{Pair shared split status}
\usage{
PairSharedSplitStatus(ref, cf)
}
\arguments{
\item{ref, cf}{Trees of class \code{\link[ape:read.tree]{phylo}} to compare.}
}
\value{
Named integer of length 6, as per \link{CompareSplits}
}
\description{
Removes all tips that do not occur in both \code{ref} and \code{cf}, then calculates
the status of the remaining splits
}
\author{
Martin R. Smith
}
\keyword{internal}
