% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PairwiseQuartets.R
\name{PairwiseQuartets}
\alias{PairwiseQuartets}
\title{Pairwise quartet distances}
\usage{
PairwiseQuartets(trees, Measure = QuartetDivergence)
}
\arguments{
\item{trees}{A list of trees of class \code{\link[ape:read.tree]{phylo}},
with identically labelled tips.}

\item{Measure}{a function that calculates tree similarity or difference
from quartet statuses. Default is \code{\link[=QuartetDivergence]{QuartetDivergence()}}.}
}
\value{
a matrix specifying the distance between each tree and
each other tree in the \code{trees}.
}
\description{
Computes the quartet distance between each pair of trees in a list.
}
\examples{
data("sq_trees")
# Calculate the status of each quartet relative to the first entry in 
# sq_trees
sq_status <- QuartetStatus(sq_trees)

# Calculate Estabrook et al's similarity measures:
SimilarityMetrics(sq_status)

# Compare trees that include a subset of the taxa 1..10
library("TreeTools", quietly = TRUE, warn.conflict = FALSE)
QuartetStatus(BalancedTree(1:5), BalancedTree(3:8), nTip = 10)

# If all taxa studied occur in `trees` or `cf`, set `nTip = TRUE`
QuartetStatus(BalancedTree(1:5), BalancedTree(3:10), nTip = TRUE)
 
}
\references{
\insertAllCited{}
}
\seealso{
\itemize{
\item Use splits (groups/clades defined by nodes or edges of the tree) instead
of quartets as the unit of comparison: \code{\link[=SplitStatus]{SplitStatus()}}.
\item Generate distance metrics from quartet statuses: \code{\link[=SimilarityMetrics]{SimilarityMetrics()}}.
}

Other element-by-element comparisons: 
\code{\link{CompareQuartets}()},
\code{\link{CompareQuartetsMulti}()},
\code{\link{CompareSplits}()},
\code{\link{PairSharedQuartetStatus}()},
\code{\link{QuartetState}()},
\code{\link{SharedQuartetStatus}()},
\code{\link{SplitStatus}()}
}
\author{
\href{https://orcid.org/0000-0001-5660-1727}{Martin R. Smith}
(\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
\concept{element-by-element comparisons}
