% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuartetDistance.R
\name{QuartetState}
\alias{QuartetState}
\title{Quartet State
State of quartets}
\usage{
QuartetState(tips, bips)
}
\arguments{
\item{tips}{A four-element array listing a quartet of tips.}

\item{bips}{bipartitions to evaluate.


 One of the three possible four-taxon trees will be consistent with any set of bipartitions 
 generated from a fully resolved tree.  If the taxa are numbered 1 to 4, this tree can be 
 identified by naming the sister taxon to taxon 1.
 If a set of bipartitions is generated from a tree that contains polytomies, it is possible
 that all three three four-taxon trees are consistent with the set of bipartitions.}
}
\value{
Returns 0 if the relationships of the four taxa are not constrained by the provided 
bipartitions, or the index of the closest relative to tips[1], otherwise.
}
\description{
Report the status of a given quartet.
}
\examples{
{
  n_tip <- 6
  trees <- list(ape::rtree(n_tip, tip.label=seq_len(n_tip), br=NULL),
                ape::rtree(n_tip, tip.label=seq_len(n_tip), br=NULL))
  splits <- lapply(trees, Tree2Splits)
  QuartetState(c(1, 3, 4, 6), splits[[2]])  
  QuartetState(1:4, splits[[1]]) == QuartetState(1:4, splits[[2]])
  vapply(Choices(n_tip), QuartetState, bips=splits[[1]], double(1))
}
}
\seealso{
\code{\link{CompareQuartets}}, used to compare sets of bipartitions
}
\author{
Martin R. Smith
}
