% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuartetDistance.R
\name{QuartetStates}
\alias{QuartetStates}
\title{describeIn QuartetStates A wrapper that need only be provided with a list of splits}
\usage{
QuartetStates(splits)
}
\arguments{
\item{splits}{a list of bipartition splits, perhaps generated using
\code{\link{Tree2Splits}}.}
}
\description{
describeIn QuartetStates A wrapper that need only be provided with a list of splits
}
\author{
Martin R. Smith
}
