% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PartitionDistance.R
\name{SplitStatus}
\alias{SplitStatus}
\alias{BipartitionStatus}
\alias{SharedSplitStatus}
\alias{SharedBipartitionStatus}
\title{Matching partitions}
\usage{
SplitStatus(trees, cf = trees[[1]])

SharedSplitStatus(trees, cf = trees[[1]])
}
\arguments{
\item{trees}{A list of trees of class \code{\link[ape:read.tree]{phylo}},
with identically labelled tips.}

\item{cf}{Comparison tree of class \code{\link[ape:read.tree]{phylo}}.  If unspecified,
each tree is compared to the first tree in \code{trees}.}
}
\value{
Returns a two dimensional array.
Rows correspond to the input trees.
Columns report the number of partitions that:
\enumerate{
\item are present in the comparison tree and the corresponding input tree;
\item are unresolved in (at least) one of the comparison tree and the corresponding
input tree.
}
}
\description{
Calculates how many of the partitions present in tree A are also present in
tree B, how many of the partitions in tree A are absent in tree B, and how
many of the partitions in tree B are absent in tree A.  The Robinson-Foulds
(symmetric partition) distance is the sum of the latter two quantities.
}
\section{Functions}{
\itemize{
\item \code{SharedSplitStatus}: Reports split statistics obtained after removing all
tips that do not occur in both trees being compared.
}}

\examples{
{
  data('sq_trees')
  
  # Calculate the status of each quartet
  splitStatuses <- SplitStatus(sq_trees)
  
  # Calculate the Robinson Foulds distances
  RobinsonFoulds(splitStatuses)
  
  # Normalize the Robinson Foulds distance by dividing by the number of 
  # splits (bipartitions) present in the two trees:
  RobinsonFoulds(splitStatuses) / splitStatuses[, 'N']
  
  # Normalize the Robinson Foulds distance by dividing by the total number of 
  # splits (bipartitions) that it is possible to resolve for `n` tips:
  nTip <- length(sq_trees[[1]]$tip.label)
  nPartitions <- 2 * (nTip - 3L) # Does not include the nTip partitions that 
                                 # comprise but a single tip
  RobinsonFoulds(splitStatuses) / nPartitions
  
}

}
\references{
{
\insertRef{Robinson1981}{Quartet}

\insertRef{Penny1985}{Quartet}
}
}
\seealso{
\itemize{
\item \code{[QuartetStatus]}: Uses quartets rather than bipartition splits as the unit
of similarity.
}
}
\author{
Martin R. Smith
}
