% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sq_trees}
\alias{sq_trees}
\title{Eighteen example trees}
\format{
An object of class \code{multiPhylo} of length 18.
}
\usage{
sq_trees
}
\description{
A list of class \code{\link[ape:multiphylo]{multiPhylo}} containing phylogenetic trees:
\describe{
\item{\code{ref_tree}}{A reference tree, bearing tips labelled 1 to 11.}
\item{\code{move_one_near}}{Tip 1 has been moved a short distance.}
\item{\code{move_one_mid}}{Tip 1 has been moved further.}
\item{\code{move_one_far}}{Tip 1 has been moved further still.}
\item{\code{move_two_near}}{Tips 10 & 11 have been moved a short distance.}
\item{\code{move_two_mid}}{Tips 10 & 11 have been moved further.}
\item{\code{move_two_far}}{Tips 10 & 11 have been moved further still.}
\item{\code{collapse_one}}{One node has been collapsed into a polytomy.}
\item{\code{collapse_some}}{Several nodes have been collapsed.}
\item{\code{m1mid_col1}}{Tree \code{move_one_mid} with one node collapsed.}
\item{\code{m1mid_colsome}}{Tree \code{move_one_mid} with several nodes collapsed.}
\item{\code{m2mid_col1}}{Tree \code{move_two_mid} with one node collapsed.}
\item{\code{m2mid_colsome}}{Tree \code{move_two_mid} with several nodes collapsed.}
\item{\code{opposite_tree}}{A tree that shares fewer quartets with \code{ref_tree} than expected by chance.}
\item{\code{caterpillar}}{A pectinate 'caterpillar' tree.}
\item{\code{top_and_tail}}{Tree \code{caterpillar}, with its outermost taxa swapped
such that it shares no partitions with \code{caterpillar}.}
\item{\code{anti_pectinate}}{A random tree that shares no partitions with
\code{caterpillar}.}
\item{\code{random_tree}}{A random tree.}
}
}
\keyword{datasets}
