% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metrics.R
\name{StatusToMatrix}
\alias{StatusToMatrix}
\title{Status vector to matrix}
\usage{
StatusToMatrix(statusVector)
}
\arguments{
\item{statusVector}{Either (i) a named vector of integers, with
names \code{N}, \code{s}, \code{r1}, \code{r2}, either \code{d} or \code{d1} and \code{d2}, and optionally \code{u}; or
(ii) a matrix whose named rows correspond to the same quantities.}
}
\value{
A matrix, containing the input columns plus \code{2d}, representing
either \code{2 * d} or \code{d1 + d2}, and row names.

The row name means that column names are dropped in
the output of \code{DoNotConflict} etc.
}
\description{
Converts a vector to a matrix that can be analysed by the \link{DoNotConflict}
function family.
}
\author{
Martin R. Smith
}
\keyword{internal}
