% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TernaryPoints.R
\name{SymmetricDifferenceLineEnds}
\alias{SymmetricDifferenceLineEnds}
\alias{SymmetricDifferenceLines}
\title{Plot lines of equal Symmetric Difference on a ternary plot}
\usage{
SymmetricDifferenceLineEnds(nsd)

SymmetricDifferenceLines(nsd, ...)
}
\arguments{
\item{nsd}{Vector specifying normalized symmetric differences to plot}

\item{\dots}{Further parameters to pass to
\code{\link[Ternary:AddToTernary]{TernaryLines}}}
}
\value{
Returns a matrix of dim \code{(length(nsd), 6)``, with columns named }r2a\code{, }da\code{, }sa\code{, }r2b\code{, }db\code{ans}sb\code{.  Lines from }a\code{to}b` in each row connect points
of equal symmetric difference.
}
\description{
Assumes that tree 1 is perfectly resolved, but that the resolution
of tree 2 can vary.
}
\section{Functions}{
\itemize{
\item \code{SymmetricDifferenceLines}: Plot the lines onto the active ternary plot
}}

\author{
Martin R. Smith
}
