% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tqDist.r
\name{TQFile}
\alias{TQFile}
\title{tqDist file generator}
\usage{
TQFile(treeList)
}
\value{
\code{TQDist()} returns a character specifying the name of the
created file.
}
\description{
Creates a temporary file corresponding to a list of trees,
to be processed with tqDist.  Files should be destroyed using
\code{on.exit(unlink(fileName))} by the calling function.
}
\details{
Used for testing purposes.
}
\keyword{internal}
