% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tqDist.r
\name{ValidateQuartetFile}
\alias{ValidateQuartetFile}
\title{Validate filenames}
\usage{
ValidateQuartetFile(file)
}
\arguments{
\item{file}{Variable to validate}
}
\value{
\code{TRUE} if \code{file} is a character vector of length one describing
a file that exists, a fatal error otherwise.
}
\description{
Verifies that file parameters are character strings describing files that exist
}
\author{
\href{https://orcid.org/0000-0001-5660-1727}{Martin R. Smith}
(\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
\keyword{internal}
