% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuartetDistance.R
\name{AllQuartets}
\alias{AllQuartets}
\title{All quartets}
\usage{
AllQuartets(n_tips)
}
\arguments{
\item{n_tips}{Integer, specifying the number of tips in a tree.}
}
\value{
Returns a list of length \code{choose(n_tips, 4)}, with each entry
corresponding to a unique selection of four different integers less than
or equal to \code{n_tips}
}
\description{
List all choices of four taxa from a tree.

A more computationally efficient alternative to \code{\link{combn}}.
Uses \code{\link{memoise}} to make repeated calls faster.
}
\examples{
 n_tips <- 6
 choice_list <- AllQuartets(n_tips)
 choice_list
 combn(n_tips, 4) # Provides the same information, but for large 
                  # values of n_tips is significantly slower.

}
\seealso{
\code{\link{combn}}
}
\author{
Martin R. Smith
}
