% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_funcs.R
\name{Choices}
\alias{Choices}
\title{Column Sums
An accellerated version of the R function \code{colSums(x, na.rm = FALSE, dims = 1L)}.
Using this function makes \code{MatchingQuartets} 8% faster.
But it is \emph{very} naughty to call \code{.Internal}, so I use the 
internal R colSums function instead.}
\usage{
Choices(n_tips)
}
\arguments{
\item{n_tips}{Integer, specifying the number of tips in a tree.}

\item{x}{Matrix whose columns are to be summed.}

\item{n_cols}{Number of columns in said matrix.}
}
\value{
Returns a list of length \code{choose(n_tips, 4)}, with each entry corresponding to a unique 
selection of four different integers <= n_tips
}
\description{
Column Sums
An accellerated version of the R function \code{colSums(x, na.rm = FALSE, dims = 1L)}.
Using this function makes \code{MatchingQuartets} 8% faster.
But it is \emph{very} naughty to call \code{.Internal}, so I use the 
internal R colSums function instead.
}
\examples{
{
 n_tips <- 6
 choice_list <- Choices(n_tips)
 choice_list
 combn(n_tips, 4) # Provides the same information, but for large values of n_tips is significantly slower.
}
}
\author{
Martin R. Smith

Martin R. Smith
}
\keyword{A}
\keyword{Choices}
\keyword{List}
\keyword{Uses}
\keyword{\code{\link{combn}}.}
\keyword{\code{\link{memoise}}}
\keyword{a}
\keyword{all}
\keyword{alternative}
\keyword{calls}
\keyword{choices}
\keyword{computationally}
\keyword{efficient}
\keyword{faster.}
\keyword{four}
\keyword{from}
\keyword{internal}
\keyword{make}
\keyword{more}
\keyword{of}
\keyword{repeated}
\keyword{taxa}
\keyword{to}
\keyword{tree.}
