% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuartetDistance.R
\name{CompareQuartets}
\alias{CompareQuartets}
\title{Compare Quartets}
\usage{
CompareQuartets(x, cf)
}
\arguments{
\item{x}{A list of quartet states, perhaps generated in
\code{\link{CompareQuartets}}.}

\item{cf}{a second such list.

Compares each quartet in a list, calculating how many statements are identical
in both lists.}
}
\value{
Returns an array of six numberic elements, each corresponding to the quantities of
Estabrook \emph{et al}. (1985):
}
\description{
Compare quartet states between trees

\describe{
  \item{Q}{The total number of quartets for _n_ tips.}
  \item{s}{The number of quartets that are resolved identically in both trees.}
  \item{d}{The number of quartets that are resolved differently in each tree.}
  \item{r1}{The number of quartets that are resolved in tree 1, but not in tree 2.}
  \item{r2}{The number of quartets that are resolved in tree 2, but not in tree 1.}
  \item{u}{The number of quartets that are unresolved in both trees.}
}
}
\details{
Compares two lists of quartet states, detailing how many are identical and
how many are unresolved.
}
\examples{
{
  n_tip <- 6
  trees <- list(ape::rtree(n_tip, tip.label=seq_len(n_tip), br=NULL),
                ape::rtree(n_tip, tip.label=seq_len(n_tip), br=NULL))
  splits <- lapply(trees, Tree2Splits)
  quartets <- QuartetStates(splits)
  compare_result <- CompareQuartets(quartets[[1]], quartets[[2]])
  dissimilar_quartets <- choose(n_tip, 4) - sum(compare_result)  
  result <- c(compare_result, dissimilar_quartets)
  names(result) <- c('Shared', 'Unresolved', 'Dissimilar')
  result
}

}
\references{
{
\insertRef{Estabrook1985}{SlowQuartet}
}
}
\seealso{
\code{\link{MatchingQuartets}}, generates this output from a list of
trees.
}
\author{
Martin R. Smith
}
