% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpdt.R
\name{CPDTDist}
\alias{CPDTDist}
\title{Wrapper for CPDT-dist}
\usage{
CPDTDist(tree1, tree2)
}
\arguments{
\item{tree1, tree2}{Phylogenetic tree of class \code{phylo}.}
}
\value{
\code{CPDTDist()} returns the quartet distance between \code{tree1} and
\code{tree2}.
}
\description{
Convenience function that passes a pair of trees to a C++ implementation
of the centroid path decomposition tree distance algorithm
\insertCite{Janssen2017jcb}{Quartet}.
}
\examples{
tree1 <- TreeTools::RootTree(TreeTools::BalancedTree(8), 1)
tree2 <- TreeTools::PectinateTree(8)
CPDTDist(tree1, tree2)

}
\references{
\insertAllCited{}
}
\author{
\href{https://orcid.org/0000-0001-5660-1727}{Martin R. Smith}
(\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})

C++ code written by Ramesh Rajaby.
}
