% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_funcs.R
\name{MatchingQuartets}
\alias{MatchingQuartets}
\title{Matching Quartets
Count matching quartets
Determines the number of four-taxon trees consistent with multiple cladograms}
\usage{
MatchingQuartets(trees)
}
\arguments{
\item{trees}{A list of trees of class \code{\link[ape]{phylo}}, with identically-labelled tips.}
}
\value{
Returns a two dimensional array; columns correspond to the input trees;
      rows report the number of four-taxon trees that : 1, are present in 
      \code{trees[[1]]} and the corresponding input tree;
      2: are unresolved in (at least) one of trees[[1]] and the corresponding 
      input tree. Tetrads that DIFFER between the two relevant trees can be 
      calculated by deducting the quartets in either of the other two
      categories from the total number of quartets, given by
       \code{\link{choose}(n_tip, 4)}.
      
      A random pair of fully-resolved trees is expected to share 
      \code{choose(n_tip, 4) / 3} four-taxon trees.
}
\description{
Given a list of trees, returns the nubmer of quartet statements present in the first tree
in the list also present in each other tree.
}
\details{
If all trees are bifurcating and the package `rtqdist` is installed, then the
function will use the faster `rtqdist`` package to generate quartet distances.
Otherwise the distances will be generated (slowly) within R.

rtqDist can be installed by copying the following code into your console:
\code{install.packages(
'http://users-cs.au.dk/cstorm/software/tqdist/files/tqDist-1.0.0.tar.gz',
 repos=NULL, type='source')}
 Or by [downloading the package](http://users-cs.au.dk/cstorm/software/tqdist/)
 and extracting the zipped directory into your library directory (which you 
 can locate by typing `.libPaths()` into your console).
  
  At present the trees must bear the same number of tips, and each tree$tip.label must use
  the integers 1:n_tip.  Support for different-sized trees will be added if there is demand; 
  contact the author if you would appreciate this functionality.
}
\examples{
{
 n_tip <- 6
 trees <- lapply(1:12, function (x) ape::rtree(n_tip, tip.label=seq_len(n_tip), br=NULL))
 compare_result <- MatchingQuartets(trees)
 dissimilar_quartets <- choose(n_tip, 4) - colSums(compare_result)  
 result <- rbind(compare_result, dissimilar_quartets)
 rownames(result) <- c('Shared', 'Unresolved', 'Dissimilar')
 result
}

}
\references{
{
\cite{SAND, A., HOLT, M. K., JOHANSEN, J., BRODAL, G. S., MAILUND, T. and 
PEDERSEN, C. N. S. 2014. 
tqDist: a library for computing the quartet and triplet distances between 
binary or general trees. \emph{Bioinformatics}, 30, 2079–2080.
<doi:10.1093/bioinformatics/btu157>
}
}
}
\author{
Martin R. Smith
}
