% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuartetDistance.R
\name{MatchingQuartets}
\alias{MatchingQuartets}
\title{Matching Quartets}
\usage{
MatchingQuartets(trees, cf = NULL)
}
\arguments{
\item{trees}{A list of trees of class \code{\link[ape:read.tree]{phylo}}, with identically-labelled tips.}

\item{cf}{Comparison tree of class \code{\link[ape:read.tree]{phylo}}.  If unspecified,
trees are compared to the first tree in \code{trees}.}
}
\value{
Returns a two dimensional array. Columns correspond to the input trees; the first column will always         report a perfect match as it compares the first tree to itself.         Rows list the status of each quartet:
\describe{
\item{Q}{The total number of quartets for \emph{n} tips.}
\item{s}{The number of quartets that are resolved identically in both trees.}
\item{d}{The number of quartets that are resolved differently in each tree.}
\item{r1}{The number of quartets that are resolved in tree 1, but not in tree 2.}
\item{r2}{The number of quartets that are resolved in tree 2, but not in tree 1.}
\item{u}{The number of quartets that are unresolved in both trees.}
}
}
\description{
Counts matching quartets
}
\details{
Determines the number of four-taxon trees consistent with multiple cladograms

Given a list of trees, returns the number of quartet statements present in the first tree
in the list also present in each other tree.

If all trees are bifurcating, then the function will use the faster
\code{tqdist} algorithm to generate quartet distances.
Otherwise the distances will be generated (slowly) by explicit enumeration.
#'
At present the trees must bear the same number of tips, and each tree$tip.label must use
the integers 1:n_tip.  Support for different-sized trees will be added if there is demand;
contact the author if you would appreciate this functionality.

A random pair of fully-resolved trees is expected to share
\code{choose(n_tip, 4) / 3} four-taxon trees.
}
\examples{
{
 n_tip <- 6
 data(sq_trees)
 qt <- MatchingQuartets(sq_trees)

 # Calculate Estabrook et al's similarity measures:
 do_not_conflict = qt[]
}

}
\references{
\insertRef{Estabrook1985}{Quartet}
\insertRef{Sand2014}{Quartet}
}
\seealso{
\link{MatchingSplits}
}
\author{
Martin R. Smith
}
