% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuartetDistance.R
\name{MatchingQuartets}
\alias{MatchingQuartets}
\title{Matching Quartets}
\usage{
MatchingQuartets(trees, use.tqDist = TRUE)
}
\arguments{
\item{trees}{A list of trees of class \code{\link[ape]{phylo}}, with identically-labelled tips.}

\item{use.tqDist}{Logical specifying whether to attempt to use the tqDist algorithm.
Requires that the \code{rtqdist} package is installed.}
}
\value{
Returns a two dimensional array.
Columns correspond to the input trees; the first column will always
report a perfect match as it compares the first tree to itself.
Rows list the status of each quartet:
}
\description{
Count matching quartets

\describe{
  \item{Q}{The total number of quartets for _n_ tips.}
  \item{s}{The number of quartets that are resolved identically in both trees.}
  \item{d}{The number of quartets that are resolved differently in each tree.}
  \item{r1}{The number of quartets that are resolved in tree 1, but not in tree 2.}
  \item{r2}{The number of quartets that are resolved in tree 2, but not in tree 1.}
  \item{u}{The number of quartets that are unresolved in both trees.}
}
}
\details{
Determines the number of four-taxon trees consistent with multiple cladograms

Given a list of trees, returns the number of quartet statements present in the first tree
in the list also present in each other tree.

If all trees are bifurcating and the package \code{rtqdist} (Sand 2014) is installed, then the
function will use the faster `rtqdist`` package to generate quartet distances.
Otherwise the distances will be generated (slowly) within R.

rtqDist can be installed by copying the following code into your console:
\code{install.packages(
'http://users-cs.au.dk/cstorm/software/tqdist/files/tqDist-1.0.0.tar.gz',
 repos=NULL, type='source')}
Or by \href{http://users-cs.au.dk/cstorm/software/tqdist/}{downloading the package}
and extracting the zipped directory into your library directory (which you
can locate by typing \code{.libPaths()} into your console).

At present the trees must bear the same number of tips, and each tree$tip.label must use
the integers 1:n_tip.  Support for different-sized trees will be added if there is demand;
contact the author if you would appreciate this functionality.\preformatted{  A random pair of fully-resolved trees is expected to share 
  \code{choose(n_tip, 4) / 3} four-taxon trees.
}
}
\examples{
{
 n_tip <- 6
 data(sq_trees)
 qt <- MatchingQuartets(sq_trees)

 # Calculate Estabrook et al's similarity measures:
 do_not_conflict = qt[]
}

}
\references{
\insertRef{Estabrook1985}{SlowQuartet}
\insertRef{Sand2014}{SlowQuartet}
}
\author{
Martin R. Smith
}
