% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuartetDistance.R
\name{PlotQuartet}
\alias{PlotQuartet}
\title{Plot quartet on a tree topology}
\usage{
PlotQuartet(tree, quartet, overwritePar = TRUE, caption = TRUE, ...)
}
\arguments{
\item{tree}{A tree of class \code{phylo}, or a list of such trees.}

\item{quartet}{A vector of four integers, corresponding to numbered tips on
the tree; or a character vector specifying the labels of four
tips.}

\item{overwritePar}{Logical specifying whether to use existing
\code{\link[graphics]{par}} \code{mfrow} and \code{mar} parameters
(\code{FALSE}),
or to plot trees side-by-side in a new graphical device (\code{TRUE}).}

\item{caption}{Logical specifying whether to annotate each plot to specify
whether the quartet selected is in the same or a different state to the
reference tree.}

\item{\dots}{Additional parameters to send to \code{\link[graphics]{plot}}}
}
\value{
Returns \code{invisible()}, having plotted a tree in which the first two members
of \code{quartet} are highlighted in orange, and the second two highlighted in
blue.
}
\description{
Draws a tree, highlighting the members of a specified quartet in colour.
}
\examples{
  data('sq_trees')
  
  par(mfrow=c(3, 5), mar=rep(0.5, 4))
  PlotQuartet(sq_trees, c(2, 5, 3, 8), overwritePar = FALSE)

}
\author{
Martin R. Smith
}
