% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TernaryPoints.R
\name{QuartetPoints}
\alias{QuartetPoints}
\alias{SplitPoints}
\alias{BipartitionPoints}
\title{Plot Tree Differences on Ternary Plots}
\usage{
QuartetPoints(trees, cf = trees[[1]])

SplitPoints(trees, cf = trees[[1]])

BipartitionPoints(trees, cf = trees[[1]])
}
\arguments{
\item{trees}{A list of trees of class \code{\link[ape:read.tree]{phylo}}, with identically-labelled tips.}

\item{cf}{Comparison tree of class \code{\link[ape:read.tree]{phylo}}.  If unspecified,
trees are compared to the first tree in \code{trees}.}
}
\value{
A data frame listing the ternary coordinates of trees, based on the
amount of information that they have in common with the comparison
tree (which defaults to the first member of the list, if unspecified).\preformatted{    If the comparison tree `cf` contains polytomies, then it is possible
    that trees within `trees` will be more resolved than `cf`.  A split
    that is resolved in `trees[[i]]` but not in `cf` will contribute 
    `+1` to the `Contradicted` count, and `-1` to the `Unresolved` count.
}
}
\description{
Generate points to depict tree difference (in terms of resolution
and similarity) on a ternary plot.
}
\section{Functions}{
\itemize{
\item \code{SplitPoints}: Uses partition distance instead of quartet metric.
}}

\examples{
{
  library('Ternary')
  data('sq_trees')
  
  TernaryPlot('Consistent', 'Contradicted', 'Unresolved')
  TernaryLines(list(c(1/3, 2/3, 0), c(0, 0, 1)), col='red', lty='dotted')
  TernaryPoints(QuartetPoints(sq_trees), col=Ternary::cbPalette8[2])
  TernaryPoints(SplitPoints(sq_trees), col=Ternary::cbPalette8[3])
}

}
\references{
\insertRef{Smith2019}{Quartet}
}
\author{
Martin R. Smith
}
