% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tqDist.r
\name{TQFile}
\alias{TQFile}
\title{tqDist file generator}
\usage{
TQFile(treeList)
}
\value{
Name of the created file
}
\description{
Creates a temporary file corresponding to a list of trees,
to be processed with tqDist.  Files should be destroyed using
\code{on.exit(file.remove(fileName))} by the calling function.
}
\details{
Should now only be necessary for testing purposes.
}
\keyword{internal}
